+++ components/filter-dropdown.vue
@@ -15,19 +15,7 @@
         aria-label="Filter"
       >
         <span class="u-icon-untouchable c-button--center-icon">
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="24"
-            height="24"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            stroke-width="2"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-          >
-            <polygon points="22 3 2 3 10 12.46 10 19 14 21 14 12.46 22 3" />
-          </svg>
+          <FilterIcon />
         </span>
         alphabetisch
       </button>
@@ -44,13 +32,14 @@
 </template>
 
 <script>
-import { X } from 'lucide-vue'
+import { X, Filter as FilterIcon } from 'lucide-vue'
 
 export default {
   name: 'FilterDropdown',
 
   components: {
-    X
+    X,
+    FilterIcon
   },
 
   data () {